add chatbot ui to docs (#845)
### Description
Adds a chatbot UI to our documentation:

![image](https://github.com/user-attachments/assets/974da52e-fdab-4d05-af67-e2d3ea296c00)


In particular, I've injected the compiled static assets for the
production build of the [NVES chat bot
app](https://gitlab-master.nvidia.com/nves/llm_bot/-/tree/main/chat_frontend)
into our `mkdocs` page. (The BioNeMo 2 docs are being indexed for this
chatbot).

Because of the different rendering strategies used between react and
mkdocs, we need to turn off the `navigation.instant` and
`navigation.instant.prefetch` features of mkdocs to allow for the chat
to work on navigation.

Note - because we don't have a great system for testing changes in live
docs, I've made the default behavior for the chatbot
invisible/unclickable, so we can turn it on with dev tools and
investigate in 'secret' 🥷.

### Type of changes
<!-- Mark the relevant option with an [x] -->

- [ ]  Bug fix (non-breaking change which fixes an issue)
- [x]  New feature (non-breaking change which adds functionality)
- [ ]  Refactor
- [x]  Documentation update
- [ ]  Other (please describe):

> [!NOTE]
> By default, the notebooks validation tests are skipped unless
explicitly enabled.


### Usage
Checkout this branch, and in the root of the repo, build the docs
```
docker build -t nvcr.io/nvidian/cvai_bnmo_trng/bionemo2-docs -f docs/Dockerfile .
```
then run the docs:

```
docker run --rm -it -p 8000:8000 \
    -v ${PWD}/docs:/docs -v ${PWD}/sub-packages:/sub-packages \
    nvcr.io/nvidian/cvai_bnmo_trng/bionemo2-docs:latest
```

---------

Signed-off-by: Jared Wilber <jwilber@nvidia.com>