run mdformat and bump ruff pre-commit (#1036)
We use mdformat to automatically format md files in bionemo-recipes, as
part of the prep work to merging these repos we could add mdformatting
to the existing bionemo-framework repo.

This adds 
```
  - repo: https://github.com/executablebooks/mdformat
    rev: 0.7.17             # Use the latest stable version
    hooks:
      - id: mdformat
        additional_dependencies:
        - mdformat-tables
        - mdformat-gfm
        - mdformat-black
        - mdformat-frontmatter
```

to .pre-commit-config.yaml, which automatically formats markdown files
(and code blocks within markdown files). It also adds an
`.mdformat.yaml` file to customize the mdformat behavior.

Signed-off-by: Peter St. John <pstjohn@nvidia.com>