Add bionemo-recipes (#1052)
Adds BioNeMo-recipes, examples for training high-performance biology
models outside of NeMo using TransformerEngine and nvFSDP.

Most of the new work in this PR is around getting CI running on the
nv_gha_runner backend. In summary, I've renamed the previous CI to
"unit-tests-framework.yml", and use step_security/changed_files to
detect when any code changes outside of recipes/ and models/

The new recipes CI lives in "unit-tests-recipes.yml", and also uses
step_security/changed_files to build a list of directories to test.
Since the models and recipes are independent folders, we only run CI for
each if a file has changed in them. We then use a matrix strategy job to
kick off a job for each changed models/ or recipes/ folder.

---------

Signed-off-by: Peter St. John <pstjohn@nvidia.com>