Remove index.md markdown to prevent 3.13 requirement (#1076)
The cards in the docs `index.md` require a specific format.
However, `mdformat` changes this format. To workaround this, we had to
add excludes functionalty for `mdformat` that prevented it from
formatting `index.md`. For some reason, the excludes functionality in
`mdformat` which requires python 3.13. This is broke our nightly build.

To fix this, I've converted the `index.md` card logic from markdown to
html, so that it no longer gets formatted. This removes the need for the
excludes functionality and the 3.13 dependency it brings.

Signed-off-by: Jared Wilber <jwilber@nvidia.com>