pin torch version in amplify (#1112)
For some reason a dependency seems to be requesting a torch update in
the AMPLIFY model tests, which we can't do in the older torch 2.6.0 base
container (to continue to support the legacy xformers version of the
original AMPLIFY model). This pins the torch version in these tests to
make sure we don't update via pip

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- Chores
  - Updated core ML runtime to a pinned Torch pre-release version.
  - Enabled Transformer Engine (PyTorch) dependency.
- Note: You may need to reinstall/update dependencies after pulling this
change.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Signed-off-by: Peter St. John <pstjohn@nvidia.com>