bump recipes container to 25.10 (#1317)
Bumps the recipes container to 25.10.

We have to xfail / disable a couple of features to get the
nvidia-internal torch fork to work:
* megatron-fsdp's `gather_uneven_dtensor_to_full_tensor` seems to break
with newer versions of torch, fix WIP.
* torchdata's StatefulDataLoader uses internal pytorch APIs that changed
with pytorch/pytorch#163102 (merged in the nvidia NGC images), so to use
StatefulDataLoader we need to set `pin_memory=False`, which impacts
performance. We now have the option to fall back to the standard pytorch
dataloader with `use_stateful_dataloader=False`

---------

Signed-off-by: Peter St. John <pstjohn@nvidia.com>