Ensure parameters are initialized correctly on the meta device (#1368)
It looks like `reset_parameters` isn't enough to ensure that
newly-initialized models using the torch.device("meta") semantics are
identical to those created without it. This PR adds tests to ensure that
these parameter mean and stdevs are identical, and uses
```python
    if args.use_meta_device:
        model.to_empty(device=device)
        model.apply(model._init_weights)
```
to ensure the weights are correctly initialized following `to_empty()`.

---------

Signed-off-by: Peter St. John <pstjohn@nvidia.com>